Changes to the book.
Made it so when you hover over the "Legendary" Status dragonfly, the information page about it pops up and disappears, when you stop hovering over it. I can make it like that for every rarity, but we only have the information page about the legendary one. There are still things that need to be polished about this, for example the button is still active even after you flip pages. Ill try to work on it tomorrow.