Blood sweat and tears.
Made it so the button changes from the silhouette to the colored image, when the bug is caught. Made a code, which keeps track of what bug has been caught , and it allows us to add new bugs and the code should still work, if we name the bugs and their corresponding images correctly. Made it so the Info page also changes from the silhouette to the colored version.  This should also be repeatable with new bug types.