Camera position, Net position, shadows,book buttons,  and preparing the game for the special bugs.
Made the camera be a bit higher up, so the player does not seem so short,  adjusted the net position a bit, so the its not inside the capsule, made that the book buttons are on-click. Had to redo the preparation for the special bugs task, because of the GitHub issues that I had, but it's done.