Butterflies.
Set up the butterfly models, AI move Waypoints and Roombas from analyzing how Rachel had set it all up for the Dragonflies. Cannot really test if it works or not since the dragonflies themselves are not working because of the recent merge. (At least that's what I think so.) Everything is done in my scene because I want to make sure that this works before I implement it into the main scene.