Lots of things.
Added butterflies +their movement to the game using Maddies made code. Fixed player capsule position,  gold Dragonfly collider, slowed down the player and made it so the movement in the cave is the exact same as outside it (just edited Maddies code's values). Took off the shadows for the player capsule and the net.